框架基本重构完成
