+++ framework_common/manshuo_draw/core/classic_collection/basic_img_set.py
@@ -1,4 +1,4 @@
-from PIL import Image, ImageDraw
+from PIL import Image, ImageDraw, ImageFilter, ImageOps
 from .initialize import initialize_yaml_must_require
 
 class basicimgset:
@@ -32,3 +32,50 @@ class basicimgset:
         return canvas
 
 
+    def combine_layer_basic(self,basic_img,layer_img_canvas):
+        width, height = layer_img_canvas.size
+        if height > self.img_height - self.padding_up_common * 2:
+            height = self.img_height - self.padding_up_common * 2
+            layer_img_canvas = layer_img_canvas.crop((0, 0, width, height))
+        # 圆角处理
+        if self.is_rounded_corners_front and self.is_rounded_corners_layer:
+            mask = Image.new("L", layer_img_canvas.size, 0)
+            draw = ImageDraw.Draw(mask)
+            draw.rounded_rectangle((0, 0, width, height), radius=self.rounded_corners_radius, fill=255, outline=255, width=2)
+            rounded_image = Image.new("RGBA", layer_img_canvas.size)
+            rounded_image.paste(layer_img_canvas, (0, 0), mask=mask)
+            layer_img_canvas=rounded_image
+        # 阴影处理
+        if self.is_shadow_front and self.is_shadow_layer:
+            shadow_image = Image.new("RGBA", basic_img.size, (0, 0, 0, 0))  # 初始化透明图层
+            shadow_draw = ImageDraw.Draw(shadow_image)
+            # 计算阴影矩形的位置
+            shadow_rect = [
+                self.padding_left_common - self.shadow_offset,  # 左
+                self.padding_up_common - self.shadow_offset,  # 上
+                self.padding_left_common + width + self.shadow_offset,  # 右
+                self.padding_up_common + height + self.shadow_offset  # 下
+            ]
+            # 绘制阴影（半透明黑色）
+            shadow_draw.rectangle(shadow_rect, fill=(0, 0, 0, self.shadow_opacity))
+            # 对阴影层应用模糊效果
+            shadow_image = shadow_image.filter(ImageFilter.GaussianBlur(self.blur_radius))
+            # 将阴影层与底层图像 layer2 合并
+            combined_shadow = Image.alpha_composite(basic_img, shadow_image)
+            combined_shadow.paste(layer_img_canvas, (self.padding_left_common,self.padding_up_common), mask=layer_img_canvas)
+            basic_img = combined_shadow
+
+        # 描边处理
+        if self.is_stroke_front and self.is_stroke_layer:
+            shadow_image = Image.new('RGBA', (width + self.stroke_layer_width, height + self.stroke_layer_width),(255, 255, 255, 80))
+            shadow_blurred = shadow_image.filter(ImageFilter.GaussianBlur(self.stroke_layer_width / 2))
+            mask = Image.new('L', shadow_blurred.size, 255)
+            draw = ImageDraw.Draw(mask)
+            draw.rounded_rectangle([0, 0, shadow_blurred.size[0], shadow_blurred.size[1]],radius=self.stroke_layer_radius, fill=0, outline=255, width=2)
+            shadow_blurred = ImageOps.fit(shadow_blurred, mask.size, method=0, bleed=0.0, centering=(0.5, 0.5))
+            mask = ImageOps.invert(mask)
+            shadow_blurred.putalpha(mask)
+            basic_img.paste(shadow_blurred, (int(self.padding_left_common - self.stroke_layer_width / 2), int(self.padding_up_common - self.stroke_layer_width / 2)), shadow_blurred.split()[3])
+
+        basic_img.paste(layer_img_canvas, (self.padding_left_common, self.padding_up_common,self.padding_left_common + width,self.padding_up_common + height), mask=layer_img_canvas)
+        return basic_img
