+++ framework_common/manshuo_draw/core/classic_collection/img.py
@@ -30,7 +30,9 @@ class ImageModule:
         for content in self.img:
             if isinstance(content, str) and os.path.splitext(content)[1].lower() in [".jpg", ".png", ".jpeg", '.webp']: #若图片为本地文件，则转化为img对象
                 self.processed_img.append(Image.open(content))
-
+            elif isinstance(content, str) and content.startswith("http"):
+                img_path=download_img(content, self.img_path_save + "/" + random_str() + ".png")
+                self.processed_img.append(Image.open(img_path))
 
             else:#最后判断是否为base64，若不是，则不添加本次图像
                 try:
@@ -41,3 +43,33 @@ class ImageModule:
 
     def common(self):
         pure_backdrop = Image.new("RGBA", (self.img_width, self.img_height), (0, 0, 0, 0))
+
+        if self.number_per_row == 'default' :
+            if len(self.processed_img) == 1:
+                self.number_per_row=1
+                self.is_crop = False
+            elif len(self.processed_img) in [2,4] : self.number_per_row=2
+            else: self.number_per_row=3
+
+        new_width = (((self.img_width - self.padding*2 ) - (self.number_per_row - 1) * self.padding_with) // self.number_per_row)
+        x_offset = self.padding
+        current_y= 0
+        number_count = 0
+        for img in self.processed_img:
+            number_count+=1
+            img = img.resize((new_width, int(new_width * img.height / img.width)))
+
+            #img = add_rounded_corners(img, radius=20)
+            #creat_white_corners(canvas, new_width, int(new_width * img.height / img.width), x_offset, current_y)
+
+            # 检查透明通道
+            if img.mode == "RGBA":pure_backdrop.paste(img, (int(x_offset), int(current_y)), img.split()[3])
+            else:pure_backdrop.paste(img, (int(x_offset), int(current_y)))
+            x_offset += new_width + self.padding_with
+            if number_count == self.number_per_row:
+                number_count = 0
+                current_y += img.height + self.padding_with
+        if number_count != 0:
+            current_y  +=  new_width * img.height / img.width
+
+        return {'canvas': pure_backdrop, 'canvas_bottom': current_y - self.padding_with}
